src/main/java/microservice/app/socialmultiplication/service/MultiplicationServiceImpl.java
package microservice.app.socialmultiplication.service;

import microservice.app.socialmultiplication.domain.Multiplication;
import microservice.app.socialmultiplication.domain.MultiplicationResultAttempt;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
class MultiplicationServiceImpl implements MultiplicationService {
    private RandomGeneratorService randomGeneratorService;

    @Autowired
    public MultiplicationServiceImpl(RandomGeneratorService randomGeneratorService) {
        this.randomGeneratorService = randomGeneratorService;
    }



    @Override
    public Multiplication createRandomMultiplication() {
        int factorA = randomGeneratorService.generateRandomFactor();
        int factorB = randomGeneratorService.generateRandomFactor();
        return new Multiplication(factorA,factorB);
    }

    @Override
    public boolean checkAttempt(MultiplicationResultAttempt resultAttempt) {
       int MultiplicationResult = resultAttempt.getMultiplication().getResult();
       return MultiplicationResult == resultAttempt.getResultAttempt();
    }
}
